server and config files may be located in subdirectory
This modifications allow to locate the configuration file in any
existent subdirectory of the output directory and adding the reference
to it from app script properly to its location.

Moreover, the `server.path` configuration parameter is used to know
where the server script is located and therefore add the reference to
it, in the app script according their locations.