Fixes Issue #18
When mimosa-config.coffee (or .js) contains the web-package module, but
no server config section (or a server config section with no path:
entry), then the crash would occur.

Added code to config.coffee to set the config.webPackage.appjs to
undefined in this case so that the code generation step is bypassed.

Also handles the case where you have a server: section in your
mimosa-config.coffee file, but no path: entry.